Fix return code of caml_minisat_value
The type of lbool is char, and the type of the return code is int.
Converting such char to int includes sign extension only on x86.
On all other architectures the return value of (char)-1 becomes 0xff
instead of the expected (int)-1. As a result all usage of Minisat.value
will trigger its assert.

Convert from char to int manually to provide the expected values.

Fixes #7

Signed-off-by: Olaf Hering <olaf@aepfle.de>